Fix accessing empty optional.
The source class accesed empty optional representing the source file.
That lead to undefined behaviour, mostly in assert crashes. This patch
fixes it and adds new tests for the offending code that caused crashes.

+++ t86/debugger/Source/Expression.cpp
@@ -233,7 +233,7 @@ void ExpressionEvaluator::Visit(const ArrayAccess& id) {
         if (idx == nullptr) {
             throw DebuggerError("Can only indexate with integers");
         }
-        if (idx->value >= av->members.size()) {
+        if (idx->value >= static_cast<int64_t>(av->members.size())) {
             throw DebuggerError(fmt::format("Out of bounds access: {} >= {}", idx->value, av->members.size()));
         }
         visitor_value = av->members[idx->value];
